refactor(api): extract chapter endpoints from books.py into chapters.py (#94)
First slice of the books.py decomposition (plan 010). Behavior-preserving
move: relocates the 9 chapter-management handlers (create/get/update/delete/
list/metadata/bulk-status/tab-state) verbatim from books.py into a new
chapters.py router, mounted under the same /books prefix so all URLs, schemas,
and status codes are unchanged.

- books.py: 3,496 to 2,938 lines (-558); 43 to 34 handlers; trimmed 10 imports
  orphaned by the move (7 chapter schemas + 3 chapter transactions)
- chapters.py: new router, 9 handlers, self-contained (no shared local helpers)
- router.py: include chapters.router with prefix="/books"
- plans/010-books-decomposition-map.md: full handler-by-cluster map (deliverable)
  with the suggested extraction order for the remaining slices
- test_chapters_characterization.py: route-table identity net - asserts every
  chapter path is still served, now owned by chapters.router, with literal
  sub-paths ordered before /{chapter_id}; existing CRUD/content suites remain
  the behavioral net (pass before + after)

Route table (method, path) verified byte-identical before/after. Full backend
suite: 961 passed / 15 skipped, 92.38% coverage.
