+++ tasks/todo.md
@@ -1,40 +1,51 @@
-# Issue #50 — P2.10 Accessibility (WCAG 2.1 AA)
+# Issue #94 (P3.1) — Decompose books.py: map + extract chapters slice
 
-Branch: `feature/50-accessibility`. Frontend-only. Driven by the real axe audit + explorer evidence, not blind edits. Solid foundation already exists (Radix, jest-axe 0 violations on 6 components, correct form labels).
+**Branch**: `feature/issue-94-decompose-books-chapters`
+**Type**: tech-debt / behavior-preserving move. NOT a redesign.
 
-## Scope: concrete, automatable, testable deliverables (mapped to ACs)
+## Context (drift re-derived)
+- `books.py` is now **3,496 lines** (plan was written at 3,184; drift confirmed, ranges re-derived).
+- 43 route handlers, no module-level helper functions (every `async def` is a route handler).
+- The chapter-management cluster is a **contiguous block, lines 1440–1994**
+  (`# Individual Chapter CRUD Operations` → end of `save_tab_state`), fully self-contained
+  (uses only module imports; no shared local helpers).
 
-### A. ARIA labels on icon-only buttons (AC: all interactive elements have ARIA labels)
-- [ ] `components/chapters/EditorToolbar.tsx` — 13 buttons have `title=` but NO `aria-label` (titles aren't read by SRs). Add `aria-label` to each.
-- [ ] `components/chapters/TabBar.tsx` — scroll up/down buttons: add `aria-label`.
-- [ ] `components/chapters/TabOverflowMenu.tsx` — trigger: `aria-label="Show more tabs"`.
-- [ ] `components/ui/user-button.tsx` — avatar trigger: `aria-label="Account menu"`.
-- (Forms already labeled — sign-in/up, BookMetadataForm verified correct → no change.)
+## Chapters slice = 9 handlers moving to `chapters.py`
+create_chapter, get_chapters_metadata, get_tab_state, get_chapter, update_chapter,
+delete_chapter, list_chapters, update_chapter_status_bulk, save_tab_state.
+(Content/analytics/batch-content, questions, drafts stay — each a future slice.)
 
-### B. Keyboard navigation (AC: full keyboard nav — tab/arrow/enter/escape)
-- [x] ~~Arrow/Home/End roving tab navigation~~ — **DEFERRED (documented)**: `TabBar` renders tabs via react-beautiful-dnd `Draggable`s, which reserve Space/arrow keys for keyboard drag-reorder. Roving arrow nav would conflict and risk breaking accessible reordering. Tabs are already keyboard-operable (each `tabIndex=0`, Enter/Space activates → WCAG 2.1.1 met) + Ctrl+1-9 quick-switch. Arrow roving is an ARIA-APG recommendation, not an AA criterion.
-- [ ] `components/books/DeleteBookModal.tsx` — wrap confirm input + actions in `<form onSubmit>` so Enter submits when confirmation matches.
+## Steps
+- [x] Step 1 — Decomposition map -> `plans/010-books-decomposition-map.md` (deliverable)
+- [x] Step 2 — Route-table snapshot before change (`/tmp/routes_before.txt`, 69 routes)
+- [x] Step 3 — Characterization safety net (`test_chapters_characterization.py` route-table identity + existing CRUD/content suites)
+- [x] Step 4 — Extracted `chapters.py`; removed moved handlers from `books.py` (3496→2938); trimmed 10 orphaned imports; wired `router.py`
+- [x] Step 5 — Routes byte-identical (69, diff clean); full suite **961 passed / 15 skipped, 92.38% cov**; net-new ruff issues cleared
 
-### C. Landmarks + skip link (AC: skip links available)
-- [ ] `app/layout.tsx` — add a `Skip to main content` link (`sr-only focus:not-sr-only`) at the top; change the wrapper `<main>` to a `<div>` (it's a layout wrapper, and nesting banner/contentinfo inside `main` is an axe violation).
-- [ ] `app/dashboard/layout.tsx` — convert content `<div className="flex-1 bg-gray-950">` to `<main id="main-content">`; add `aria-label` to the `<nav>`s.
-- [ ] `app/page.tsx` + auth `sign-in`/`sign-up` — ensure a single `<main id="main-content">` landmark wraps page content (skip-link target on public pages).
+## Result
+- `books.py`: 3,496 → 2,938 lines (−558); 43 → 34 handlers
+- `chapters.py`: new, 9 handlers, 87% covered
+- Route table (method, path) **byte-identical** before/after
 
-### D. Focus indicators (AC: focus indicators clearly visible)
-- [ ] `app/globals.css` — add a global `*:focus-visible` outline fallback (complements component rings) for elements without their own ring.
+## Acceptance criteria (from issue)
+- [ ] `plans/010-books-decomposition-map.md` catalogs all handlers by cluster + line range
+- [ ] `chapters.py` holds the chapter handlers; `books.py` no longer contains them
+- [ ] (method, path) route set byte-identical before/after
+- [ ] Characterization tests pass before AND after extraction
+- [ ] `books.py` line count meaningfully reduced (report before/after)
+- [ ] `ruff check app/` exit 0; `pytest tests/` no NEW failures
+- [ ] No URL/schema/status-code changes; no out-of-scope files touched
 
-### E. Color contrast (AC: contrast meets AA) — narrow; dark theme mostly passes
-- [ ] `components/SummaryInput.tsx:104` `text-gray-500 dark:text-gray-600` help text → bump dark to `gray-400` (gray-600 on dark fails AA). Spot-fix only genuine failures; do NOT mass-swap (most `gray-400` on dark passes ~6:1).
-
-### F. Automated a11y tests (AC: automated a11y tests pass)
-- [ ] Extend `src/__tests__/accessibility/ComponentAccessibilityAudit.test.tsx` — unskip/ add Navigation (dashboard layout), EditorToolbar, and a SkipLink/landmark test (jest-axe → 0 violations).
-- [ ] Unit tests: EditorToolbar aria-labels, ChapterTabs arrow/Home/End nav, DeleteBookModal Enter-submits, skip-link present + targets `#main-content`.
-
-## ACs that are manual/deploy-time (documented, not auto-verifiable here)
-- Lighthouse a11y >95 — deploy/manual (changes only improve it).
-- Screen reader (NVDA/JAWS) announcements — manual; we wire the correct ARIA (role=alert/aria-live already documented + used).
-
-## Gates
-- [ ] lint + typecheck clean; full unit suite + coverage (≥85/85/75/85) green.
-- [ ] jest-axe suite green (existing + new).
-- [ ] codex cross-family review clean.
+## Autonomous decisions (no architectural fork)
+1. **Characterization tests**: an equivalent behavioral suite already exists
+   (`test_books_chapters_crud_coverage.py` + `test_books_chapter_content_coverage.py`
+   cover CRUD/metadata/bulk-status/tab-state, happy/auth/not-owned/404). Rather than
+   duplicate ~740 lines, add a lean `test_chapters_characterization.py` whose unique
+   value is the **route-table identity** assertion (every chapter path still served
+   after the move — the one invariant specific to this refactor) plus tab-state smoke
+   coverage. Existing suites remain the behavioral net (pass before + after).
+2. **Test file location**: `tests/test_api/test_routes/` (where all chapter tests live),
+   not the plan's literal `tests/test_api/` — sibling convention wins.
+3. **Mount**: `router.include_router(chapters.router, prefix="/books", tags=["chapters"])`
+   after the books include; ordering within `chapters.py` preserves metadata/tab-state
+   before `{chapter_id}` so no route shadowing.
